backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.15
websockets==12.0

